+++ src/Ydb.Sdk/src/Ado/Session/PoolingSessionSource.cs
@@ -23,6 +23,7 @@ internal sealed class PoolingSessionSource<T> : ISessionSource where T : Pooling
     private readonly TimeSpan _sessionIdleTimeout;
     private readonly Timer _cleanerTimer;
     private readonly ILogger _logger;
+    private readonly string? _frameworkClientInfo;
 
     private volatile int _numSessions;
     private volatile int _disposed;
@@ -46,6 +47,7 @@ internal sealed class PoolingSessionSource<T> : ISessionSource where T : Pooling
         _sessionIdleTimeout = TimeSpan.FromSeconds(settings.SessionIdleTimeout);
         _cleanerTimer = new Timer(CleanIdleSessions, this, _sessionIdleTimeout, _sessionIdleTimeout);
         _logger = settings.LoggerFactory.CreateLogger<PoolingSessionSource<T>>();
+        _frameworkClientInfo = settings.ClientInfo;
 
         MetricsReporter = new YdbMetricsReporter(_maxPoolSize, _minPoolSize, Statistics, settings);
     }
@@ -322,6 +324,12 @@ internal sealed class PoolingSessionSource<T> : ISessionSource where T : Pooling
         catch (Exception e)
         {
             _logger.LogError(e, "Failed to dispose the transport driver");
+        }
+
+        SdkClientInfoRegistry.Unregister($"ado-net/{YdbSdkVersion.Value}");
+        if (_frameworkClientInfo is not null)
+        {
+            SdkClientInfoRegistry.Unregister(_frameworkClientInfo);
         }
 
         for (var i = 0; i < _maxPoolSize; i++)