+++ src/Ydb.Sdk/src/Metadata.cs
@@ -19,4 +19,7 @@ internal static class Metadata
 
     //Incoming hints
     public const string GracefulShutdownHint = "session-close";
+
+    internal static readonly string TopicWriterClientInfo = $"topic-writer/{YdbSdkVersion.Value}";
+    internal static readonly string TopicReaderClientInfo = $"topic-reader/{YdbSdkVersion.Value}";
 }